% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcasm.R
\name{randcase}
\alias{randcase}
\title{Randomly select \code{toupper} or \code{tolower}}
\usage{
randcase()
}
\value{
A function, either the definition of \code{toupper} or \code{tolower}
}
\description{
Randomly select \code{toupper} or \code{tolower}
}
