% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcasm.R
\name{sarcastic}
\alias{sarcastic}
\title{Generate sarcastic version of text input}
\usage{
sarcastic(text)
}
\arguments{
\item{text}{A character vector, of length >= 1}
}
\value{
A character vector of length equal to \code{text}.
}
\description{
Generate sarcastic version of text input
}
