% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcasm.R
\name{sarcastic_errors}
\alias{sarcastic_errors}
\title{Print error messages sarcastically}
\usage{
sarcastic_errors(silent = FALSE)
}
\arguments{
\item{silent}{Boolean, prints reset instructions if TRUE}
}
\value{
Nothing, used for its side effect in setting options for errors.
}
\description{
Errors not snarky enough for you? Print them snarkily. Not recommended for
non-interactive use.
}
