% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcasm.R
\name{/s}
\alias{/s}
\title{Shorthand version of function sarcasm}
\usage{
text \%>\% `/s`
}
\arguments{
\item{text}{A character vector, of length >= 1}
}
\value{
A character vector of length equal to \code{text}.
}
\description{
Reddit-style sarcasm. Most useful after a pipe.
}
